Converts challenge admin widget and refactors managehacc tests
