+++ app/imports/ui/components/administrator/ChallengesAdminWidget.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Button, Table } from 'semantic-ui-react';
+import { Button } from 'react-bootstrap';
 import PropTypes from 'prop-types';
 import swal from 'sweetalert';
 import { Link } from 'react-router-dom';
@@ -8,8 +8,8 @@ import { Challenges } from '../../../api/challenge/ChallengeCollection';
 import { COMPONENT_IDS } from '../../testIDs/componentIDs';
 
 /** Renders a single row in the table. See pages/Listmenuitemss.jsx. */
-class ChallengesAdminWidget extends React.Component {
-  removeItem(docID) {
+const ChallengesAdminWidget = ({ challenges }) => {
+  const removeItem = (docID) => {
     swal({
       title: 'Are you sure?',
       text: 'Once deleted, you will not be able to recover this challenge!',
@@ -29,34 +29,27 @@ class ChallengesAdminWidget extends React.Component {
             swal('You canceled the deletion!');
           }
         });
-  }
+  };
 
-  render() {
-    return (
-        <Table.Row>
-          <Table.Cell width={2}>{this.props.challenges.title}</Table.Cell>
-          <Table.Cell width={5}>{this.props.challenges.description}</Table.Cell>
-          <Table.Cell width={2}>{this.props.challenges.submissionDetail}</Table.Cell>
-          <Table.Cell width={2}>{this.props.challenges.pitch}</Table.Cell>
-          <Table.Cell width={2}>
-            <Button>
-              <Link to={`/edit-challenge/${this.props.challenges._id}`}
-                    style={{ color: 'rgba(0, 0, 0, 0.6)' }}
-                    id={COMPONENT_IDS.EDIT_CHALLENGE_BUTTON}
-              >
-                Edit
-              </Link>
-            </Button>
-          </Table.Cell>
-          <Table.Cell width={2}>
-            <Button negative onClick={() => this.removeItem(this.props.challenges._id)}>
-              Delete
-            </Button>
-          </Table.Cell>
-        </Table.Row>
-    );
-  }
-}
+  return (
+      <tr>
+        <td>{challenges.title}</td>
+        <td>{challenges.description}</td>
+        <td>{challenges.submissionDetail}</td>
+        <td>{challenges.pitch}</td>
+        <td>
+          <Button variant="outline-primary">
+            <Link to={`/edit-challenge/${challenges._id}`} id={COMPONENT_IDS.EDIT_CHALLENGE_BUTTON}>Edit</Link>
+          </Button>
+        </td>
+        <td>
+          <Button variant="outline-danger" onClick={() => removeItem(challenges._id)}>
+            Delete
+          </Button>
+        </td>
+      </tr>
+  );
+};
 
 /** Require a document to be passed to this component. */
 ChallengesAdminWidget.propTypes = {
